Introduce a class based approach for the CLI tool.
This means it's very easy to extend honcho as wells as maintain it.

Some features of argparse's default behaviour can be worked around,
e.g. "common arguments" are possible with the `parents` argument
when creating the subparser, but fails when displaying the usage
and help messages, e.g. the help of a subtask shows the options
and other subcommands, too.

This also adds a bit more compatibility to foreman, e.g. a help
subcommand that is able to show the help message for other
subcommands as well as the global help message.

The check subcommand now print with the honcho.command logger.

Added a common exception for errors happening during the run of a
subcommand (`CommandError`).

Adds a `--version` option to the honcho command.

Signed-off-by: Nick Stenning <nick@whiteink.com>