Fixed several tests so they pass on Windows. Only three failures left.
This is a work in progress: we are down to only 3 tests failures on
Windows. Several of the fixes are about adjusting expectations for
windows line ending support (\r\n instead of the posix \n) or
updating test fixtures to have a posix and a
windows-specific flavors of procfiles with windows commands instead of
bash so that they can run on both platforms.