Explain how honcho helps make python apps easier to get running for new developers.
I don't expect this to be merged as-is, but it seemed like a good way
to explain how I personally found honcho to be helpful. Circus has a
similar use case in that it promotes giving ops a circus config and
letting it do the rest as far as managing processes and services.

Thanks for writing honcho!