Remove install-time dependency on Jinja2
As discussed at length in #50, an install-time dependency on Jinja2,
especially when locked to a patch version of the package, causes more
problems than it solves.

This commit moves the Jinja2 dependency to a setuptools "extra", so
that `pip install honcho` no longer pulls it in, but it is still
possible to automatically install with

    pip install honcho[export]

As this is a breaking change (for people who were relying on `pip
install honcho; honcho export ...`) it will require a minor version
bump.

Fixes #50.