Check for PEP8-compatible naming
Install a flake8 plugin to check for PEP8 naming convention
compatibility.

Only one "error" in the code base that I think is an instance of
pep8-naming actually being wrong. While __new__ is not a class method,
it's also not an instance method (it's a static method) and so should
have "cls" as its first arg, not "self".