Ignore Ctrl-C in `honcho run`
When running ad hoc commands under Honcho with `honcho run`, Honcho
currently does not give the spawned process an opportunity to handle
Ctrl-C. In addition to an ugly `KeyboardInterrupt` traceback, this
can cause the terminal to be left in an inconsistent state.

For instance, when running a command that switches the terminal to
raw mode (`honcho run python`), the terminal is left in raw mode after
honcho exits, requiring an `stty sane` to restore it.

This patch causes the `honcho run` command to ignore Ctrl-C (SIGINT)
while waiting for the spawned command process to exit, allowing the
command to handle it however it sees fit.