process: Consolidate to a single printer
This commit simplifies the process by which Honcho relays process output
to another place. Previously, every Process object had its own printer
instance, which knew what colour and name to print for the process. Now,
there is just one Printer instance, which needs to know only the width
of the prefix column, and each line printed by a process is relayed to
the printer with colour and name information.