supervisord template: \n at end; not at beginning
It's a UNIX convention that text files should always end with a newline.
One of the best reasons for this is that it makes possible to do `cat
file1.txt file2.txt > combined.txt` and have the result make sense. For
more info:

http://stackoverflow.com/questions/729692/why-should-files-end-with-a-newline

I also modified the template slightly so it doesn't generate a blank
line at the beginning of the file. I don't think it does any harm, but I
think it looks slightly better without it.