export supervisord: Improve process numbering
More consistent with Foreman (see #109).

1. Make numbering 1-based instead of 0-based.
2. Include number in program name.

I also moved the append of `num` to `full_name` from the Python code to
the template. I think that this is better, because it will allow folks
to control whether or not it gets appended using a custom template
(which is hopefully coming soon... :smile:)