Replace string_escape with simple replacements
Holy mother of God. Making string_escape work across Python 2 and 3 is
an absolute nightmare. The bytes-to-bytes encoders were removed in
Python 3, so the only way to use them is to use unicode_escape directly
from the codecs package, but this makes a complete hash of the task if
the data you give it contains unescaped unicode:

    >>> codecs.decode('åéïò', 'unicode_escape')
    'Ã¥Ã©Ã¯Ã²'
    >>> codecs.decode('åéïò'.encode('utf8'), 'unicode_escape')
    'Ã¥Ã©Ã¯Ã²'

WTF!? Anyway. Life is too short. This will make the tests pass.

Fixes #128.