Fix quoting of special env var values in Upstart configs
As reported by #151, the quoting of environment variable values in
Upstart configs is currently broken in the general case.

Correctly quoting variable values in the current context (i.e. already
inside single quotes) is possible, but extremely messy, and results in
hugely nested quotes when those values contain single quotes.

Instead, move the definition of the environment variables into the body
of the Upstart config using the `env` directive, and add the `-m` option
to `su(1)` so that environment variables are passed through to the shell
process.

Closes #151.
Closes #154.