Don't add ANSI color sequences on non-TTY outputs
This adds a check to the Printer class so that it only adds ANSI color
codes to an output line when the output object passes a isatty() check.

This is advisable in situations where honcho's output is not written to
a terminal (e.g. redirected to a file, started as a systemd service,
started in a detached docker container, ...) and therefore would only
pollute the output target with non-interpreted ANSI escape sequences.