Flush output after each line
I've gone back and forth on this one. On the one hand, doing this could
negatively honcho's throughput. On the other, this is the behaviour that
I believe most people expect from honcho, even when piping output to a
file or in other situations where STDOUT is not a TTY.

On balance, I'd rather have the default behaviour be unsurprising, and
so this makes honcho's output line-buffered, regardless of Python's
output buffering settings.