Start to clean up config precedence
Precedence for different configuration sources (CLI, app .env files, OS
environent, defaults) was implemented differently for each configuration
variable.

This commit starts to clean this up by moving configuration into a
simple map of string to string. This map is in fact a chainmap
reflecting configuration precedence.

At the moment whether specific variables can be overridden by app config
or OS config is "opt in" and controlled by APP_ENV_ARGS and OS_ENV_ARGS
respectively. The only reason these two are not the same set is to
reflect the current implementation and not because they should really be
different. That will change in the very next commit.